+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Change Log
 v0.6 (not yet released)
 -----------------------
 
+* Print function in christmas tree pattern
+  `#62 <https://github.com/msoeken/kitty/pull/62>`_
+
 v0.5 (October 8, 2018)
 ----------------------